fix: removed all warnings, added Cowx

+++ src/FWApplication.h
@@ -9,35 +9,13 @@ struct SDL_Renderer;
 struct SDL_Window;
 struct SDL_Texture;
 typedef struct _TTF_Font TTF_Font;
-//struct SDL_Event;
-
-//#define SDL_DRAW_PUTPIXEL_BPP(A, B, C)  \
-//*(A(B(Uint8*)super->pixels + (y0+y)*super->pitch +               \
-//                                          (x0+x)*SDL_DRAW_BPP)) = C; \
-//*(A(B(Uint8*)super->pixels + (y0-y)*super->pitch +               \
-//                                          (x0+x)*SDL_DRAW_BPP)) = C; \
-//*(A(B(Uint8*)super->pixels + (y0+y)*super->pitch +               \
-//                                          (x0-x)*SDL_DRAW_BPP)) = C; \
-//*(A(B(Uint8*)super->pixels + (y0-y)*super->pitch +               \
-//                                          (x0-x)*SDL_DRAW_BPP)) = C;
-//
-//
-//#define SDL_DRAW_PUTPIXEL \
-//  SDL_DRAW_PUTPIXEL_BPP(0+,1+,colorbyte1)   \
-//  if (SDL_BYTEORDER == SDL_BIG_ENDIAN) {   \
-//    SDL_DRAW_PUTPIXEL_BPP(0+,0+,colorbyte2)   \
-//    SDL_DRAW_PUTPIXEL_BPP(0+,2+,colorbyte0) \
-//          }else{                                   \
-//    SDL_DRAW_PUTPIXEL_BPP(0+,0+,colorbyte0)   \
-//    SDL_DRAW_PUTPIXEL_BPP(0+,2+,colorbyte2) \
-//          }
 
 struct Color
 {
-	uint32_t r, g, b, a;
+	uint8_t r, g, b, a;
 
 	Color() {}
-	Color(uint32_t red, uint32_t green, uint32_t blue, uint32_t alpha) : r(red), g(green), b(blue), a(alpha) { }
+	Color(uint8_t red, uint8_t green, uint8_t blue, uint8_t alpha) : r(red), g(green), b(blue), a(alpha) { }
 };
